+++ feather-runtime/examples/server.rs
@@ -1,6 +1,6 @@
 use feather_runtime::{http, server::server::Server};
 fn main() {
-    env_logger::init();
+    simple_logger::init_with_level(log::Level::Debug);
     let config = feather_runtime::server::server::ServerConfig{
         address: "127.0.0.1:5000".to_string(),
     };
@@ -8,6 +8,6 @@ fn main() {
     println!("Server started on {}", server.address());
     server.incoming().for_each(|_r|{
         
-        return http::HttpResponse::default()
+        return http::Response::default()
     }).unwrap();
 }